Fix index controller test to show page to fill out new answer.
It block describes test to 'display page to answer survey'.
Existing route pointed to page which showed a survey summary
for the owner and had a link to a page which allowed the user
to answer the survey. I updated the route tested to be the answer
page and updated the include text to reflect text present on this
page.